cleanup: callback error returns
Not sure allowing a callback to return an error is a good idea here,
since we can't get this error back to the user.

But for now, if we allow callbacks to return errors, we might as well
log them to know what is going on. (not return/pass them through, since
this should be recoverable).